legacy-be/pyproject.toml
[tool.poetry]
name = "major-proj"
version = "0.1.0"
description = ""
authors = ["shawshankkumar <shashankkumarthakur@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.109.0"}
black = "^24.1.0"
ulid-py = "^1.1.0"
boto3 = "^1.34.29"
libsql-experimental = "^0.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
